🚚 (route): refactor component organisation
Prepare for the insertion of other interface and refactor components
organisation to distinguish the different interface.

For improved readability the new components do not include "Component"
in their names.
