+++ box-ui/src/components/css/BoxMeeting.css
@@ -1,4 +1,4 @@
-.CreateMeetingComponent {
+.BoxMeeting {
     display: flex;
     flex-direction: column;
     justify-content: flex-start;
