+++ box-ui/src/components/tests/BoxHome.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import CreateMeetingComponent from '../ts/CreateMeetingComponent';
+import BoxHome from '../ts/Box/BoxHome';
 
 test('renders learn react link', () => {
-    render(<CreateMeetingComponent />);
+    render(<BoxHome />);
     const linkElement = screen.getByText(/learn react/i);
     expect(linkElement).toBeInTheDocument();
 });
